Add new install file for each suite
* Add install file for deep
* Create installer for viz
* Mv viz tests into their own directory
* Add TODO for flags
* Add external tests
* Mv everything back out
* Add install for mc
* Rename isntall file
* Move inject in install suite
* Go back on Viz flag changes

Add -p flag to run tests sequentially
