+++ bin/_test-helpers.sh
@@ -381,7 +381,7 @@ run_test(){
 
   printf 'Test script: [%s] Params: [%s]\n' "${filename##*/}" "$*"
   # Exit on failure here
-  GO111MODULE=on go test -test.timeout=60m --failfast --mod=readonly "$filename" --linkerd="$linkerd_path" --helm-path="$helm_path" --default-allow-policy="$default_allow_policy" --k8s-context="$context" --integration-tests "$@" || exit 1
+  GO111MODULE=on go test -p 1 -test.timeout=60m --failfast --mod=readonly "$filename" --linkerd="$linkerd_path" --helm-path="$helm_path" --default-allow-policy="$default_allow_policy" --k8s-context="$context" --integration-tests "$@" || exit 1
 }
 
 # Returns the latest version for the release channel
@@ -455,7 +455,7 @@ upgrade_test() {
   )
   exit_on_err "upgrade_test() - installing viz extension in $upgrade_version failed"
 
-  run_test "$test_directory/install/install_test.go" --upgrade-from-version="$upgrade_version"
+  run_test "$test_directory/install/1_install_test.go" --upgrade-from-version="$upgrade_version"
 }
 
 # Run the upgrade-edge test by upgrading the most-recent edge release to the
@@ -467,7 +467,6 @@ run_upgrade-edge_test() {
 
 run_viz_test() {
   local tests=()
-  run_test "$test_directory/install/install_test.go" --viz
   run_test "$test_directory/viz/..."
 }
 
@@ -517,12 +516,12 @@ run_helm-upgrade_test() {
 
   setup_helm
   helm_viz_chart="$( cd "$bindir"/.. && pwd )"/viz/charts/linkerd-viz
-  run_test "$test_directory/install/install_test.go" --helm-path="$helm_path" --helm-charts="$helm_charts" \
+  run_test "$test_directory/install/1_install_test.go" --helm-path="$helm_path" --helm-charts="$helm_charts" \
   --viz-helm-chart="$helm_viz_chart" --helm-stable-chart='linkerd/linkerd2' --viz-helm-stable-chart="linkerd/linkerd-viz" --helm-release="$helm_release_name" --upgrade-helm-from-version="$stable_version"
   helm_cleanup
 }
 
-run_uninstall_test() {
+run_un1_install_test() {
   run_test "$test_directory/install/uninstall/uninstall_test.go" --uninstall=true
 }
 
@@ -533,65 +532,45 @@ run_multicluster_test() {
   "$bindir"/certs-openssl
   cd "$pwd"
   export context="k3d-target"
-  run_test "$test_directory/install/install_test.go" --multicluster --certs-path "$tmp"
-  run_test "$test_directory/multicluster/target1" --multicluster
+  run_test "$test_directory/multicluster/1_install_test.go" --certs-path "$tmp"
+  run_test "$test_directory/multicluster/target1"
   link=$(multicluster_link target)
 
   export context="k3d-source"
   # Create the emojivoto namespace in the source cluster so that mirror services
   # can be created there.
   kubectl --context="$context" create namespace emojivoto
-  run_test "$test_directory/install/install_test.go" --multicluster --viz --certs-path "$tmp"
+  run_test "$test_directory/multicluster/1_install_test.go" --certs-path "$tmp"
   echo "$link" | kubectl --context="$context" apply -f -
-  run_test "$test_directory/multicluster/source" --multicluster
-
+  run_test "$test_directory/multicluster/source" 
   export context="k3d-target"
-  run_test "$test_directory/multicluster/target2" --multicluster
-
+  run_test "$test_directory/multicluster/target2" 
   export context="k3d-target"
-  run_test "$test_directory/multicluster/target-statefulset" --multicluster
+  run_test "$test_directory/multicluster/target-statefulset" 
 }
 
 run_deep_test() {
   local tests=()
-  run_test "$test_directory/install/install_test.go"
   run_test "$test_directory/deep/..."
 }
 
 run_default-policy-deny_test() {
   local tests=()
   export default_allow_policy='deny'
-  run_test "$test_directory/install/install_test.go" --viz
+  run_test "$test_directory/install/1_install_test.go" 
 }
 
 run_cni-calico-deep_test() {
   local tests=()
-  run_test "$test_directory/install/install_test.go" --cni --calico
   run_test "$test_directory/deep/..." --cni
 }
 
-run_helm-deep_test() {
-  local tests=()
-  setup_helm
-  helm_multicluster_chart="$( cd "$bindir"/.. && pwd )"/multicluster/charts/linkerd-multicluster
-  helm_viz_chart="$( cd "$bindir"/.. && pwd )"/viz/charts/linkerd-viz
-  run_test "$test_directory/install/install_test.go" --helm-path="$helm_path" --helm-charts="$helm_charts" \
-  --helm-release="$helm_release_name" --multicluster-helm-chart="$helm_multicluster_chart" \
-  --viz-helm-chart="$helm_viz_chart" --multicluster-helm-release="$helm_multicluster_release_name"
-  while IFS= read -r line; do tests+=("$line"); done <<< "$(go list "$test_directory"/.../...)"
-  for test in "${tests[@]}"; do
-    run_test "$test"
-  done
-  helm_cleanup
-}
-
 run_external_test() {
-  run_test "$test_directory/install/install_test.go" --external-issuer=true --viz --external-prometheus=true
   run_test "$test_directory/external/..."
 }
 
 run_cluster-domain_test() {
-  run_test "$test_directory/install/install_test.go" --cluster-domain='custom.domain' --viz
+  run_test "$test_directory/install/1_install_test.go" --cluster-domain='custom.domain' 
 }
 
 # exit_on_err should be called right after a command to check the result status
