+++ test/integration/external/testdata/check.viz.external-prometheus.golden
