+++ test/integration/install/inject/inject_test.go
