+++ test/integration/install/inject/testdata/inject_test.yaml
