+++ test/integration/install/inject/testdata/pods.yaml
