+++ test/integration/install/inject/testdata/services.yaml
