+++ test/integration/install/smoke/install_smoke_test.go
