+++ test/integration/install/smoke/testdata/check.proxy.golden
