+++ test/integration/install/smoke/testdata/inject.report.golden
