+++ test/integration/install/smoke/testdata/smoke_test.yaml
