--- test/integration/install/testdata/check.cni.golden
@@ -1,78 +1,0 @@
-Linkerd core checks
-===================
-
-kubernetes-api
---------------
-√ can initialize the client
-√ can query the Kubernetes API
-
-kubernetes-version
-------------------
-√ is running the minimum Kubernetes API version
-√ is running the minimum kubectl version
-
-linkerd-existence
------------------
-√ 'linkerd-config' config map exists
-√ heartbeat ServiceAccount exist
-√ control plane replica sets are ready
-√ no unschedulable pods
-√ control plane pods are ready
-√ cluster networks contains all node podCIDRs
-
-linkerd-config
---------------
-√ control plane Namespace exists
-√ control plane ClusterRoles exist
-√ control plane ClusterRoleBindings exist
-√ control plane ServiceAccounts exist
-√ control plane CustomResourceDefinitions exist
-√ control plane MutatingWebhookConfigurations exist
-√ control plane ValidatingWebhookConfigurations exist
-√ proxy-init container runs as root user if docker container runtime is used
-
-linkerd-cni-plugin
-------------------
-√ cni plugin ConfigMap exists
-√ cni plugin ClusterRole exists
-√ cni plugin ClusterRoleBinding exists
-√ cni plugin ServiceAccount exists
-√ cni plugin DaemonSet exists
-√ cni plugin pod is running on all nodes
-
-linkerd-identity
-----------------
-√ certificate config is valid
-√ trust anchors are using supported crypto algorithm
-√ trust anchors are within their validity period
-√ trust anchors are valid for at least 60 days
-√ issuer cert is using supported crypto algorithm
-√ issuer cert is within its validity period
-√ issuer cert is valid for at least 60 days
-√ issuer cert is issued by the trust anchor
-
-linkerd-webhooks-and-apisvc-tls
--------------------------------
-√ proxy-injector webhook has valid cert
-√ proxy-injector cert is valid for at least 60 days
-√ sp-validator webhook has valid cert
-√ sp-validator cert is valid for at least 60 days
-√ policy-validator webhook has valid cert
-√ policy-validator cert is valid for at least 60 days
-
-linkerd-version
----------------
-√ can determine the latest version
-√ cli is up-to-date
-
-control-plane-version
----------------------
-√ can retrieve the control plane version
-√ control plane is up-to-date
-√ control plane and cli versions match
-
-linkerd-control-plane-proxy
----------------------------
-√ control plane proxies are healthy
-√ control plane proxies are up-to-date
-√ control plane proxies and cli versions match
