+++ test/integration/multicluster/source/source_test.go
@@ -16,10 +16,6 @@ var TestHelper *testutil.TestHelper
 
 func TestMain(m *testing.M) {
 	TestHelper = testutil.NewTestHelper()
-	if !TestHelper.Multicluster() {
-		fmt.Fprintln(os.Stderr, "Multicluster test disabled")
-		os.Exit(0)
-	}
 	os.Exit(m.Run())
 }
