+++ test/integration/multicluster/target-statefulset/target_statefulset_test.go
@@ -26,10 +26,6 @@ var (
 
 func TestMain(m *testing.M) {
 	TestHelper = testutil.NewTestHelper()
-	if !TestHelper.Multicluster() {
-		fmt.Fprintln(os.Stderr, "Multicluster test disabled")
-		os.Exit(0)
-	}
 	os.Exit(m.Run())
 }
