+++ test/integration/multicluster/target1/target_test.go
@@ -2,7 +2,6 @@ package target1
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"testing"
 
@@ -13,10 +12,6 @@ var TestHelper *testutil.TestHelper
 
 func TestMain(m *testing.M) {
 	TestHelper = testutil.NewTestHelper()
-	if !TestHelper.Multicluster() {
-		fmt.Fprintln(os.Stderr, "Multicluster test disabled")
-		os.Exit(0)
-	}
 	os.Exit(m.Run())
 }
