+++ test/integration/multicluster/target2/target_test.go
@@ -14,10 +14,6 @@ var TestHelper *testutil.TestHelper
 
 func TestMain(m *testing.M) {
 	TestHelper = testutil.NewTestHelper()
-	if !TestHelper.Multicluster() {
-		fmt.Fprintln(os.Stderr, "Multicluster test disabled")
-		os.Exit(0)
-	}
 	os.Exit(m.Run())
 }
