+++ test/integration/viz/testdata/check.viz.golden
